create folder for nav, separate displays
