+++ client/joinus_client/src/components/Nav/NavDisplayUser.js
@@ -1,0 +1,56 @@
+import {
+  Avatar,
+  IconButton,
+  Menu,
+  MenuItem,
+  Tooltip,
+  Typography,
+} from "@mui/material";
+import { Box } from "@mui/system";
+
+const settings = ["Events", "Settings", "Logout"];
+
+export default function NavDisplayUser(props) {
+  const {
+    user,
+    handleOpenUserMenu,
+    anchorElUser,
+    handleCloseUserMenu,
+    handleLogout,
+  } = props;
+
+  return (
+    <Box sx={{ flexGrow: 0 }}>
+      <Tooltip title="Open settings">
+        <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
+          <Avatar alt={user.name} src={user.picture} />
+        </IconButton>
+      </Tooltip>
+      <Menu
+        sx={{ mt: "45px" }}
+        id="menu-appbar"
+        anchorEl={anchorElUser}
+        anchorOrigin={{
+          vertical: "top",
+          horizontal: "right",
+        }}
+        keepMounted
+        transformOrigin={{
+          vertical: "top",
+          horizontal: "right",
+        }}
+        open={Boolean(anchorElUser)}
+        onClose={handleCloseUserMenu}
+      >
+        {settings.map((setting) => (
+          <MenuItem
+            key={setting}
+            onClick={setting === "Logout" ? handleLogout : handleCloseUserMenu}
+          >
+            <Typography textAlign="center">{setting}</Typography>
+          </MenuItem>
+        ))}
+      </Menu>
+    </Box>
+  );
+}
